+++ lib/dom_tools.dart
@@ -2,3 +2,4 @@
 library dom_tools;
 
 export 'src/dom_tools_base.dart';
+export 'src/dom_tools_paint.dart';