+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.1
+
+- CSS functions: applyCSS(), defineCSS(), hasCSS()
+- Class Color (from dart:ui).
+- Class CanvasImageViewer.
+
 ## 1.0.0
 
 - Initial version, created by Stagehand